Added Sticky Header functionality. Added an DividerItemDecorator into library. Moved into common folder the SmoothScrollLinearLayoutManager class. Fixed others minor bugs.

+++ flexible-adapter-app/src/main/java/eu/davidea/utils/Utils.java
@@ -48,7 +48,7 @@ public final class Utils {
 		return context.getResources().getDisplayMetrics();
 	}
 
-	public static int dpToPx(Context context, float dp) {
+	public static float dpToPx(Context context, float dp) {
 		final float scale = context.getResources().getDisplayMetrics().density;
 		return Math.round(dp * scale);
 	}
